Created: print_functions.c
- Add: pall function
The opcode pall prints all the values on the stack, starting
from the top of the stack.

- Add: pint function
The opcode pint prints the value at the top of the stack,
followed by a new line. If the stack is empty, print the error
message 'L<line_number>: can't pint, stack empty',
followed by a new line, and exit with the status EXIT_FAILURE.